requirements.txt
numpy
pandas
scipy
python-dateutil